Return Not Supported for both nfs4_op_delegpurge and CLAIM_DELEGATE_PREV
Currnt code mostly supports CLAIM_DELEGATE_PREV but there is lot more
needed to fully support it.

Some of the missing pieces are:
1. Fully supporting nfs4_op_delegpurge.
2. making sure that the delegations are not removed/cleaned on
   setclientid sequence.

This is helpful for those clients which uses persistent storage
for data cache, and these delegations are called client-persistent delegations
When there is needed we can address above mentioned issues and enable then.

Signed-off-by: Venkateswararao Jujjuri (JV) <jvrao@linux.vnet.ibm.com>